Address review: type scan's console getDsoInfo, drop getDsoInfoViaScan helper
DsoInfo and its decoder now live in HttpScanAppClient (scan is the canonical
server of the shared GetDsoInfoResponse schema); the scan console command
decodes via the console environment's template decoder, so the preflight
tests' bespoke ResourceTemplateDecoder helper is gone and scan clients are
used directly. The raw GetDsoInfo command is kept for ScanConnection
(BFT response comparison and the validator scan-proxy need the undecoded
response).

Verified: apps-scan/apps-sv/apps-app-Test compile and ScanIntegrationTest
(incl. the retyped scan-vs-SV dso-info comparison) against local Canton,
7/7 passed.
