+++ apps/app/src/test/scala/org/lfdecentralizedtrust/splice/integration/plugins/EventHistorySanityCheckPlugin.scala
@@ -88,7 +88,8 @@ class EventHistorySanityCheckPlugin(
                     .fromJson(exercised.choiceArgument.noSpaces)
                     .newSvParty == otherScan
                     .getDsoInfo()
-                    .svPartyId
+                    .svParty
+                    .toProtoPrimitive
                 case _ => false
               })
           }
