+++ apps/app/src/test/scala/org/lfdecentralizedtrust/splice/integration/tests/runbook/RunbookSvPreflightIntegrationTestBase.scala
@@ -78,7 +78,7 @@ abstract class RunbookSvPreflightIntegrationTestBase
   "The SV rewards are claimed by the SV, with 33.33% going to validator1" in { implicit env =>
     val sv1ScanClient = scancl("sv1Scan")
 
-    val dsoInfo = getDsoInfoViaScan(scancl("svTestScan"))
+    val dsoInfo = scancl("svTestScan").getDsoInfo()
     val svParty = dsoInfo.svParty.toProtoPrimitive
     val svInfo = dsoInfo.dsoRules.payload.svs.asScala.get(svParty).value
     val joinedAsOfRound = svInfo.joinedAsOfRound.number
